ReplaceAllUsesOfValueWith may delete other nodes that the one being replaced. Do not delete dead nodes again.
llvm-svn: 92988

+++ llvm/lib/CodeGen/SelectionDAG/SelectionDAGISel.cpp
@@ -459,7 +459,8 @@ void SelectionDAGISel::ShrinkDemandedOps() {
     SDNode *N = Worklist.pop_back_val();
 
     if (N->use_empty() && N != CurDAG->getRoot().getNode()) {
-      CurDAG->DeleteNode(N);
+      if (N->getOpcode() != ISD::DELETED_NODE)
+        CurDAG->DeleteNode(N);
       continue;
     }
